+++ changelog.md
@@ -1,6 +1,12 @@
+# 0.8.4
+- add check for lowercase characters
+- support wide range of uppercase unicode characters during javascript validation  
+
 # 0.8.3
 - use innodb as db engine
+
 # 0.8.2
 - update license to GPL 3
+
 # 0.8.1
 - added composer.json